# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LeetCoach AI API",
    "description": "AI coding mentor for LeetCode \u2014 hints, patterns, complexity, debugging, review, and revision tracking.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/problems": {
      "post": {
        "tags": [
          "problems"
        ],
        "summary": "Upsert Problem",
        "description": "Register/extract a problem the user opened. Also records an attempt.",
        "operationId": "upsert_problem_api_v1_problems_post",
        "parameters": [
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProblemUpsert"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "problems"
        ],
        "summary": "List Problems",
        "description": "List problems in the registry (optionally filtered).",
        "operationId": "list_problems_api_v1_problems_get",
        "parameters": [
          {
            "name": "difficulty",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Difficulty"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          },
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProblemOut"
                  },
                  "title": "Response List Problems Api V1 Problems Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/problems/{slug}/attempts": {
      "post": {
        "tags": [
          "problems"
        ],
        "summary": "Create Attempt",
        "description": "Record a submission outcome (accepted/wrong/attempted) for a problem.",
        "operationId": "create_attempt_api_v1_problems__slug__attempts_post",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProblemUpsert"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/problems/{slug}": {
      "get": {
        "tags": [
          "problems"
        ],
        "summary": "Get Problem",
        "operationId": "get_problem_api_v1_problems__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/revisions": {
      "get": {
        "tags": [
          "revisions"
        ],
        "summary": "List Revisions",
        "operationId": "list_revisions_api_v1_revisions_get",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RevisionOut"
                  },
                  "title": "Response List Revisions Api V1 Revisions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "revisions"
        ],
        "summary": "Create Revision",
        "operationId": "create_revision_api_v1_revisions_post",
        "parameters": [
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevisionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/revisions/{revision_id}": {
      "delete": {
        "tags": [
          "revisions"
        ],
        "summary": "Delete Revision",
        "operationId": "delete_revision_api_v1_revisions__revision_id__delete",
        "parameters": [
          {
            "name": "revision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Revision Id"
            }
          },
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Revision Api V1 Revisions  Revision Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AttemptOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "time_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Ms"
          },
          "first_try": {
            "type": "boolean",
            "title": "First Try"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "language",
          "time_ms",
          "first_try",
          "tags",
          "created_at"
        ],
        "title": "AttemptOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProblemContext": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug",
            "default": ""
          },
          "leetcode_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Leetcode Id"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "difficulty": {
            "type": "string",
            "title": "Difficulty",
            "default": "Medium"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "examples": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Examples"
          },
          "constraints": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Constraints"
          },
          "function_signature": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Function Signature"
          },
          "url": {
            "type": "string",
            "title": "Url",
            "default": ""
          },
          "code": {
            "type": "string",
            "title": "Code",
            "default": ""
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "python"
          }
        },
        "type": "object",
        "title": "ProblemContext",
        "description": "Normalised problem payload extracted by the extension's content script."
      },
      "ProblemOut": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "leetcode_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Leetcode Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "difficulty": {
            "type": "string",
            "title": "Difficulty"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "pattern_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pattern Key"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "slug",
          "leetcode_id",
          "title",
          "difficulty",
          "tags",
          "pattern_key",
          "description",
          "url"
        ],
        "title": "ProblemOut"
      },
      "ProblemUpsert": {
        "properties": {
          "problem": {
            "$ref": "#/components/schemas/ProblemContext"
          },
          "status": {
            "type": "string",
            "pattern": "^(accepted|wrong|attempted|solved)$",
            "title": "Status"
          },
          "first_try": {
            "type": "boolean",
            "title": "First Try",
            "default": false
          },
          "time_ms": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 86400000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Ms"
          }
        },
        "type": "object",
        "required": [
          "problem",
          "status"
        ],
        "title": "ProblemUpsert"
      },
...
      "RevisionCreate": {
        "properties": {
          "problem_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Problem Slug"
          },
          "kind": {
            "type": "string",
            "pattern": "^(observation|pattern|mistake|tip)$",
            "title": "Kind"
          },
          "content": {
            "type": "string",
            "maxLength": 20000,
            "minLength": 1,
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "content"
        ],
        "title": "RevisionCreate"
      },
      "RevisionOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "problem_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Problem Slug"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "problem_slug",
          "kind",
          "content",
          "created_at"
        ],
        "title": "RevisionOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}